feat: forbid abandoning signed funding channels
It's not safe to abandon funding channels that we have sent signatures
to the peer, since the funding tx can be committed to the chain later.

A future task is planned to make the channel inactive and continue
monitoring the funding tx until it is committed or rejected because of
double-spending.